+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 6.1
 --------------
 - Add CoursePeriodDeleteSQL() & CourseDeleteSQL() functions in Courses.fnc.php
+- Add SendNotificationNewStudentAccount() & SendNotificationNewUserAccount() functions in SendNotifications.fnc.php
 
 Changes in 6.0
 --------------